refactor(api)!: move Cancellable to the event package
Move the Cancellable interface, and the AbstractCancellable abstract
implementation to the `dev.hypera.chameleon.event` package.
